Update new-term-request.md
Failure to recognize a constitutive splice site, currently it is called an "intron gain variant", which is
too vague, since it can also result from activation of cryptic splice sites as well. I suggest an alternative term, "intron inclusion" and provide numerous examples in publications and in ClinVar to describe the consequences of failing to recognize an exon due to mutations.